Split utilities crate into modules
