+++ tests/integration/merge_fail.rs
@@ -2,7 +2,7 @@ use assert_matches::assert_matches;
 use nickel_lang::error::{Error, EvalError};
 use nickel_lang::position::TermPos;
 use nickel_lang::term::RichTerm;
-use nickel_lang_utilities::TestProgram;
+use nickel_lang_utilities::test_program::TestProgram;
 use std::io::Cursor;
 
 fn eval_full(s: &str) -> Result<RichTerm, Error> {
