+++ tests/integration/pretty.rs
@@ -1,6 +1,6 @@
 use nickel_lang::pretty::*;
 use nickel_lang::term::{RichTerm, StrChunk, Term};
-use nickel_lang_utilities::parse;
+use nickel_lang_utilities::test_program::parse;
 use pretty::BoxAllocator;
 
 use std::io::{Cursor, Read};
